Addresses #4
The error seems to not be an issue with the script but the API that is
being used.

According to tvMaze forums, today they were experiencing a higher than
normal rate of DDOS attacks, which was causing such errors.

Nevertheless, this commit error checks the data fetching. If data fails
to be fetched then it will stop the script and email the logs
regardless of settings specified by the user.

The error message will be appended to the end of the email.