The cache did not seem to work before.
I added the cache to the last update to store the hash of the last hash
that was found from github to check for script updates.

Whenever the script would fetch the hash from the script from the
previous execution it would fail to get the hash for some reason.

This commit fixes it.