Script update notification is now more robust.
Instead of parsing the HTML page like before, now I am just using the
Github API which means that it shouldn't need to be changed at all or
nearly as often.

Note that the checking for updates functionality still works, but you
may need an API token in order to avoid rate limiting. Instructions are
in the README for how to do this now.